A wrapper function that executes MantaID workflow.

A wrapper function that executes MantaID workflow.

Usage

mi(
  mi_data_allID,
  cores = NULL,
  levels = c("*", 0:9, letters, LETTERS, "_", ".", "-", " ", "/", "\\", ":"),
  ratio = 0.3,
  para_blc = FALSE,
  model_path = NULL,
  batch_size = 128,
  epochs = 64,
  validation_split = 0.3,
  graph_path = NULL
)

Arguments

mi_data_allID

IDs raw data.

cores

The number of cores used when balancing data.

levels

The vector that includes all the single characters occurred in IDs.

ratio

The ratio of the test set.

para_blc

A logical value whether using parallel computing when balancing data.

model_path

The path to save models.

batch_size

The batch size of deep learning model fitting.

epochs

The epochs of deep learning model fitting.

validation_split

The validation ratio of deep learning model fitting.

graph_path

The path to save graphs.

Value

The list of models and graphs.