Data balance. Most classes adopt random undersampling, while a few classes adopt smote method to oversample to obtain relatively balanced data;

Data balance. Most classes adopt random undersampling, while a few classes adopt smote method to oversample to obtain relatively balanced data;

Usage

mi_balance_data(data, ratio = 0.3, parallel = FALSE)

Arguments

data

A data frame. Except for class column, all are numeric types.

ratio

Numeric between 0 and 1. The percent of the test set split from data.

parallel

Logical.

Value

A list contains a train set and a test set.

Examples

library(dplyr)
data <- rename(iris, class = Species)
mi_balance_data(data)
#> [[1]]
#>     Sepal.Length Sepal.Width Petal.Length Petal.Width      class
#> 1            5.1         3.5          1.4         0.2     setosa
#> 2            4.9         3.0          1.4         0.2     setosa
#> 3            4.7         3.2          1.3         0.2     setosa
#> 4            4.6         3.1          1.5         0.2     setosa
#> 5            5.4         3.9          1.7         0.4     setosa
#> 6            4.6         3.4          1.4         0.3     setosa
#> 7            5.0         3.4          1.5         0.2     setosa
#> 8            4.4         2.9          1.4         0.2     setosa
#> 9            4.9         3.1          1.5         0.1     setosa
#> 10           5.4         3.7          1.5         0.2     setosa
#> 11           4.8         3.4          1.6         0.2     setosa
#> 12           4.8         3.0          1.4         0.1     setosa
#> 13           4.3         3.0          1.1         0.1     setosa
#> 14           5.7         4.4          1.5         0.4     setosa
#> 15           5.4         3.9          1.3         0.4     setosa
#> 16           5.1         3.5          1.4         0.3     setosa
#> 17           5.1         3.8          1.5         0.3     setosa
#> 18           4.6         3.6          1.0         0.2     setosa
#> 19           5.1         3.3          1.7         0.5     setosa
#> 20           4.8         3.4          1.9         0.2     setosa
#> 21           5.0         3.0          1.6         0.2     setosa
#> 22           5.0         3.4          1.6         0.4     setosa
#> 23           5.2         3.5          1.5         0.2     setosa
#> 24           5.2         3.4          1.4         0.2     setosa
#> 25           4.7         3.2          1.6         0.2     setosa
#> 26           5.2         4.1          1.5         0.1     setosa
#> 27           5.0         3.2          1.2         0.2     setosa
#> 28           4.9         3.6          1.4         0.1     setosa
#> 29           5.1         3.4          1.5         0.2     setosa
#> 30           5.0         3.5          1.3         0.3     setosa
#> 31           4.5         2.3          1.3         0.3     setosa
#> 32           5.0         3.5          1.6         0.6     setosa
#> 33           4.8         3.0          1.4         0.3     setosa
#> 34           4.6         3.2          1.4         0.2     setosa
#> 35           5.3         3.7          1.5         0.2     setosa
#> 36           7.0         3.2          4.7         1.4 versicolor
#> 37           6.4         3.2          4.5         1.5 versicolor
#> 38           6.9         3.1          4.9         1.5 versicolor
#> 39           5.5         2.3          4.0         1.3 versicolor
#> 40           5.7         2.8          4.5         1.3 versicolor
#> 41           4.9         2.4          3.3         1.0 versicolor
#> 42           6.6         2.9          4.6         1.3 versicolor
#> 43           5.2         2.7          3.9         1.4 versicolor
#> 44           5.0         2.0          3.5         1.0 versicolor
#> 45           5.9         3.0          4.2         1.5 versicolor
#> 46           6.0         2.2          4.0         1.0 versicolor
#> 47           5.6         2.9          3.6         1.3 versicolor
#> 48           6.7         3.1          4.4         1.4 versicolor
#> 49           5.6         3.0          4.5         1.5 versicolor
#> 50           5.8         2.7          4.1         1.0 versicolor
#> 51           5.6         2.5          3.9         1.1 versicolor
#> 52           5.9         3.2          4.8         1.8 versicolor
#> 53           6.1         2.8          4.7         1.2 versicolor
#> 54           6.6         3.0          4.4         1.4 versicolor
#> 55           6.8         2.8          4.8         1.4 versicolor
#> 56           6.0         2.9          4.5         1.5 versicolor
#> 57           5.5         2.4          3.7         1.0 versicolor
#> 58           5.8         2.7          3.9         1.2 versicolor
#> 59           6.0         2.7          5.1         1.6 versicolor
#> 60           6.0         3.4          4.5         1.6 versicolor
#> 61           6.7         3.1          4.7         1.5 versicolor
#> 62           6.3         2.3          4.4         1.3 versicolor
#> 63           6.1         3.0          4.6         1.4 versicolor
#> 64           5.8         2.6          4.0         1.2 versicolor
#> 65           5.0         2.3          3.3         1.0 versicolor
#> 66           5.6         2.7          4.2         1.3 versicolor
#> 67           5.7         2.9          4.2         1.3 versicolor
#> 68           6.2         2.9          4.3         1.3 versicolor
#> 69           5.1         2.5          3.0         1.1 versicolor
#> 70           5.7         2.8          4.1         1.3 versicolor
#> 71           6.3         3.3          6.0         2.5  virginica
#> 72           5.8         2.7          5.1         1.9  virginica
#> 73           6.3         2.9          5.6         1.8  virginica
#> 74           6.5         3.0          5.8         2.2  virginica
#> 75           4.9         2.5          4.5         1.7  virginica
#> 76           7.3         2.9          6.3         1.8  virginica
#> 77           6.4         2.7          5.3         1.9  virginica
#> 78           5.7         2.5          5.0         2.0  virginica
#> 79           5.8         2.8          5.1         2.4  virginica
#> 80           6.4         3.2          5.3         2.3  virginica
#> 81           7.7         3.8          6.7         2.2  virginica
#> 82           7.7         2.6          6.9         2.3  virginica
#> 83           5.6         2.8          4.9         2.0  virginica
#> 84           7.7         2.8          6.7         2.0  virginica
#> 85           6.7         3.3          5.7         2.1  virginica
#> 86           7.2         3.2          6.0         1.8  virginica
#> 87           6.2         2.8          4.8         1.8  virginica
#> 88           6.1         3.0          4.9         1.8  virginica
#> 89           7.2         3.0          5.8         1.6  virginica
#> 90           7.4         2.8          6.1         1.9  virginica
#> 91           7.9         3.8          6.4         2.0  virginica
#> 92           6.1         2.6          5.6         1.4  virginica
#> 93           7.7         3.0          6.1         2.3  virginica
#> 94           6.4         3.1          5.5         1.8  virginica
#> 95           6.0         3.0          4.8         1.8  virginica
#> 96           6.9         3.1          5.4         2.1  virginica
#> 97           6.7         3.1          5.6         2.4  virginica
#> 98           6.9         3.1          5.1         2.3  virginica
#> 99           5.8         2.7          5.1         1.9  virginica
#> 100          6.7         3.3          5.7         2.5  virginica
#> 101          6.7         3.0          5.2         2.3  virginica
#> 102          6.3         2.5          5.0         1.9  virginica
#> 103          6.5         3.0          5.2         2.0  virginica
#> 104          6.2         3.4          5.4         2.3  virginica
#> 105          5.9         3.0          5.1         1.8  virginica
#> 
#> [[2]]
#>    Sepal.Length Sepal.Width Petal.Length Petal.Width      class
#> 1           4.9         3.0          1.4         0.2     setosa
#> 2           4.7         3.2          1.3         0.2     setosa
#> 3           4.6         3.4          1.4         0.3     setosa
#> 4           4.4         2.9          1.4         0.2     setosa
#> 5           5.1         3.5          1.4         0.3     setosa
#> 6           5.7         3.8          1.7         0.3     setosa
#> 7           5.4         3.4          1.7         0.2     setosa
#> 8           5.1         3.3          1.7         0.5     setosa
#> 9           5.0         3.4          1.6         0.4     setosa
#> 10          4.9         3.1          1.5         0.2     setosa
#> 11          4.4         3.0          1.3         0.2     setosa
#> 12          5.1         3.4          1.5         0.2     setosa
#> 13          4.5         2.3          1.3         0.3     setosa
#> 14          4.8         3.0          1.4         0.3     setosa
#> 15          5.0         3.3          1.4         0.2     setosa
#> 16          5.5         2.3          4.0         1.3 versicolor
#> 17          4.9         2.4          3.3         1.0 versicolor
#> 18          6.6         2.9          4.6         1.3 versicolor
#> 19          5.0         2.0          3.5         1.0 versicolor
#> 20          5.6         2.9          3.6         1.3 versicolor
#> 21          5.6         3.0          4.5         1.5 versicolor
#> 22          6.4         2.9          4.3         1.3 versicolor
#> 23          6.8         2.8          4.8         1.4 versicolor
#> 24          5.7         2.6          3.5         1.0 versicolor
#> 25          5.8         2.7          3.9         1.2 versicolor
#> 26          6.0         2.7          5.1         1.6 versicolor
#> 27          5.4         3.0          4.5         1.5 versicolor
#> 28          5.5         2.5          4.0         1.3 versicolor
#> 29          5.6         2.7          4.2         1.3 versicolor
#> 30          5.7         2.8          4.1         1.3 versicolor
#> 31          7.1         3.0          5.9         2.1  virginica
#> 32          7.3         2.9          6.3         1.8  virginica
#> 33          6.8         3.0          5.5         2.1  virginica
#> 34          5.8         2.8          5.1         2.4  virginica
#> 35          7.7         3.8          6.7         2.2  virginica
#> 36          6.9         3.2          5.7         2.3  virginica
#> 37          5.6         2.8          4.9         2.0  virginica
#> 38          6.4         2.8          5.6         2.1  virginica
#> 39          7.2         3.0          5.8         1.6  virginica
#> 40          7.4         2.8          6.1         1.9  virginica
#> 41          6.4         2.8          5.6         2.2  virginica
#> 42          6.1         2.6          5.6         1.4  virginica
#> 43          6.7         3.1          5.6         2.4  virginica
#> 44          6.9         3.1          5.1         2.3  virginica
#> 45          5.9         3.0          5.1         1.8  virginica
#>